+++ scheduler-service/src/main/java/com/codeaddi/scheduler_service/model/repository/sessions/SessionRepository.java
@@ -1,5 +1,6 @@
 package com.codeaddi.scheduler_service.model.repository.sessions;
 
+import com.codeaddi.scheduler_service.model.repository.sessions.entities.Session;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
