+++ scheduler-service/src/main/java/com/codeaddi/scheduler_service/model/repository/sessions/UpcomingSessionsAvailabilityRepository.java
@@ -1,0 +1,11 @@
+package com.codeaddi.scheduler_service.model.repository.sessions;
+
+import com.codeaddi.scheduler_service.model.repository.sessions.entities.UpcomingSessionAvailability;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UpcomingSessionsAvailabilityRepository extends JpaRepository<UpcomingSessionAvailability, Long> {
+
+}
+
