+++ scheduler-service/src/main/java/com/codeaddi/scheduler_service/model/repository/sessions/entities/Session.java
@@ -1,4 +1,4 @@
-package com.codeaddi.scheduler_service.model.repository.sessions;
+package com.codeaddi.scheduler_service.model.repository.sessions.entities;
 
 import com.codeaddi.scheduler_service.model.enums.RowerLevel;
 import com.codeaddi.scheduler_service.model.enums.SessionType;
