+++ scheduler-service/src/test/java/com/codeaddi/scheduler_service/controller/db/SessionsServiceTests.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
-import com.codeaddi.scheduler_service.model.repository.sessions.Session;
+import com.codeaddi.scheduler_service.model.repository.sessions.entities.Session;
 import com.codeaddi.scheduler_service.model.repository.sessions.SessionRepository;
 import com.codeaddi.scheduler_service.testUtils.TestData;
 import jakarta.persistence.EntityNotFoundException;
